rag-lit-synthesis/pyproject.toml
[project]
name = "rag-lit-synthesis"
version = "0.1.0"
description = "RAG for Scientific Literature Synthesis: Benchmark and Evaluation"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "sentence-transformers>=3.0",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.7",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "requests>=2.31",
    "tqdm>=4.66",
    "jsonlines>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
